Refactor: BaseViewModelFactory로 병합
