--- app/src/main/java/com/sopt/now/core/base/factory/AuthViewModelFactory.kt
@@ -1,21 +1,0 @@
-package com.sopt.now.core.base.factory
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.sopt.now.data.repository.AuthRepository
-import com.sopt.now.presentation.login.LoginViewModel
-import com.sopt.now.presentation.signup.SignUpViewModel
-
-class AuthViewModelFactory(private val authRepository: AuthRepository) : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(LoginViewModel::class.java)) {
-            @Suppress("UNCHECKED_CAST")
-            return LoginViewModel(authRepository) as T
-        }
-        if (modelClass.isAssignableFrom(SignUpViewModel::class.java)) {
-            @Suppress("UNCHECKED_CAST")
-            return SignUpViewModel(authRepository) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
