--- app/src/main/java/com/sopt/now/core/base/factory/FriendViewModelFactory.kt
@@ -1,16 +1,0 @@
-package com.sopt.now.core.base.factory
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.sopt.now.data.repository.FriendRepository
-import com.sopt.now.presentation.home.HomeViewModel
-
-class FriendViewModelFactory(private val friendRepository: FriendRepository) : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(HomeViewModel::class.java)) {
-            @Suppress("UNCHECKED_CAST")
-            return HomeViewModel(friendRepository) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
