--- app/src/main/java/com/sopt/now/core/base/factory/UserViewModelFactory.kt
@@ -1,16 +1,0 @@
-package com.sopt.now.core.base.factory
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.sopt.now.data.repository.UserRepository
-import com.sopt.now.presentation.main.MainViewModel
-
-class UserViewModelFactory(private val userRepository: UserRepository) : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(MainViewModel::class.java)) {
-            @Suppress("UNCHECKED_CAST")
-            return MainViewModel(userRepository) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
